refactor: ID 型を int64 に統一
- ドメインエンティティ、usecase、handler の ID 型を int64 に統一
- JWT subject を文字列で発行し、安全な範囲の旧数値形式も受理
- OpenAPI の watchlist ID を int64 として再生成し、関連テストを更新

+++ internal/feature/auth/adapters/oauth_account_repository.go
@@ -46,7 +46,7 @@ func (r *oauthAccountRepository) Create(ctx context.Context, acct *entity.OAuthA
 		return errors.New("account is nil")
 	}
 	row, err := r.q.CreateOAuthAccount(ctx, authsqlc.CreateOAuthAccountParams{
-		UserID:      int64(acct.UserID),
+		UserID:      acct.UserID,
 		Provider:    acct.Provider,
 		ProviderUid: acct.ProviderUID,
 	})
